Add partial rendering example
This takes a slightly different approach to the mixin-based one in #5 that doesn't need any new features in django-htmx.

By using Django's ability to dynamically select a base template in `{% extends %}`, we avoid a need for `{% htmx_include %}`.

By using the inheritance of htmx attributes, we avoid the need for the `htmx_attrs` tag to repeat common attributes.