Vendor htmx (#513)
I avoided vendoring htmx initially because it was moving quite fast and
I didn't want to keep updating this package with new versions. Since v2
it has settled, and so to make it easier to use htmx with Django, it’s
now vendored. Render it with a new template tag, `{% htmx_script %}`,
which also adds the django-htmx extension script when `DEBUG` is `True`.